requirements.txt
streamlit>=1.52.0
google-genai>=1.57.0
pandas>=2.2.0
pysqlite3-binary